trains works with project now. folds work as well. finished data loader fixes.
