+++ requirements.txt
@@ -7,3 +7,4 @@ numpy~=1.19.4
 scikit-image~=0.18.0
 torchvision~=0.8.2
 pandas~=1.1.5
+clearml~=0.17.2
