+++ src/configs.yml
@@ -5,8 +5,8 @@ dataset:
   img_foldername: img
 model:
   use_cuda: True
-  use_saved: False
-  path: 12_29_15-49-29_test100ep_w_good_fold
+  use_saved: True
+  path: 12_30_17-09-42_sanity-fold-full-dataset-w-aug
   weight_init: True
   name: UNet
   use_bn: True
@@ -16,24 +16,26 @@ model:
   lr: 1.e-2
 train:
   batch_size: 4
-  epochs: 100
+  epochs: 1000
   subset_size:
   use_folds: True
   val_round: True
+  shuffle_val: True
   val_percent: 0.2
-  run_name: 'shuffle_folds'
+  run_name: 'sanity-fold-full-dataset-w-aug'
   print_every: 1
   save: True
-  save_every: 10
-  data_augmentation: # TODO: data augement.
-    horizontal_flip: False
+  save_every: 50
+  data_augmentation:
+    horizontal_flip: True
     flip_p: 0.5
-    color_jitter: False
-    gaussian_blur: False
-    gaussian_noise: False
+    color_jitter: True
+    gaussian_blur: True
+    gaussian_noise: True
     # possibly add std and probability for augs over here?
 
 evaluate: # TODO: actually take eval into account.
   batch_size: 10
   hist: False
 verbose: True
+random_seed: 42
